Eseguita l'integrazione tra gaussian_elimination.c e matrix.c; Sistemato un piccolo problema di dichiarazione in matrix.c e matrix.h.

+++ include/matrix.h
@@ -8,7 +8,7 @@ void set_matrix(unsigned int** matrix, unsigned int x, unsigned int y, unsigned 
 
 void init_matrix_l(unsigned long*** matrix, unsigned int x, unsigned int y);
 void finalize_matrix_l(unsigned long*** matrix, unsigned int x);
-unsigned int get_matrix_l(unsigned long** matrix, unsigned int x, unsigned int y);
+unsigned long get_matrix_l(unsigned long** matrix, unsigned int x, unsigned int y);
 void set_matrix_l(unsigned long** matrix, unsigned int x, unsigned int y, unsigned int value);
 
 #endif // MATRIX_QUARD
